Renamed and moved the ListCard component to GenericMediumCard.vue
