+++ src/components/GenericMediumCard.vue
